pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-model-comparison"
version = "0.1.0"
description = "A modular tool for comparing AI models using Gemini and ScaleDown.ai"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "flask",
    "python-dotenv",
    "google-generativeai",
    "requests",
    "pandas"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["ai_model_comparison*"]
